pyproject.toml
[tool.poetry]
name = "foo"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.17.5"
markupsafe = "2.0.1"
gunicorn = "^20.0.4"
sqlalchemy = "^1.4.31"
noonhelpers = {url = "https://storage.googleapis.com/noon-toolbox/2019/python-noonhelpers/noonhelpers-master.tar.gz?cb=5"}
noonutil = {url = "https://storage.googleapis.com/noon-toolbox/2020/python-noonutil/noonutil-v225.tar.gz"}
eventpubsub = {url = "https://storage.googleapis.com/noon-toolbox/2018/python-eventpubsub/eventpubsub-v0.9.8.tar.gz"}
google-cloud-pubsub = "1.7.0"
requests = "^2.23.0"
sentry-sdk = "^1.5.6"
jsql = "0.7"
google-api-core = "^1.31.5"
grpcio-gcp = "0.2.2"
py-spy = "^0.3.11"
pylint = "2.7.1"
retrying = "1.3.3"
boltons = "21.0.0"
pytest = "^6.2.5"
toml = "0.10.0"
pyhumps = "^1.3.1"
async-timeout = "^3.0.1"
jinja2 = "^2.11.2"
sqlparse = "^0.3.0"
moz-sql-parser = "^4.18.21031"
google-cloud-spanner = "1.19.1"
google-cloud-core="1.7.2"
fastapi = "0.65.3"
Shapely = "1.8.0"
aiosolr = "2.5.3"
pysolr = "3.7.0"
fastkml = "0.11"
google-cloud-storage = "1.33.0"
lazy-string = "^1.0.0"
python-gettext = "^4.0"
Babel = "^2.9.1"
http3 = "0.6.7"
retrying-async = "1.2.0"
time-machine = "^2.5.0"
mysql-connector-python = "8.0.26"
google-cloud-logging = "^2.7.0"
opentelemetry-instrumentation-fastapi = "^0.26b1"
opentelemetry-sdk = "^1.7.1"
opentelemetry-exporter-gcp-trace = "^1.0.0"
opentelemetry-instrumentation-requests = "^0.26b1"
opentelemetry-instrumentation-sqlalchemy = "^0.26b1"
spanner-orm = "^0.5.0"
mysqlclient = "2.0.3" # engineutil (noonutil v225) only compqtible with this version due to: https://github.com/PyMySQL/mysqlclient/pull/513/files
google-cloud-secret-manager = "^2.9.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.12b0"

[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true
skip-magic-trailing-comma = true

[build-system]
requires = ["poetry_core>=1.1.4"]
build-backend = "poetry.core.masonry.api"

[pytest]
log_cli = "True"